+++ src/groups/membership.rs
@@ -12,11 +12,7 @@ use super::{roles, user_role};
 /// - 4: You do not have permission to manage this member.
 /// - 18: The operation is temporarily unavailable. Please try again later.
 /// - 25: 2-Step Verification is required to make further transactions. Go to Settings > Security to complete 2-Step Verification.
-pub async fn remove_user(
-    jar: &RequestJar,
-    group_id: usize,
-    user_id: usize,
-) -> Result<(), Box<Error>> {
+pub async fn remove_user(jar: &RequestJar, group_id: u32, user_id: u32) -> Result<(), Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/groups/{}/users/{}",
         group_id, user_id
@@ -28,7 +24,7 @@ pub async fn remove_user(
 #[derive(Debug, Serialize, Deserialize, Clone)]
 #[serde(rename_all = "camelCase")]
 pub struct UpdateUserRoleRequest {
-    pub role_id: usize,
+    pub role_id: u32,
 }
 
 /// Sets a user's role in a group.
@@ -43,9 +39,9 @@ pub struct UpdateUserRoleRequest {
 /// - 26: You cannot change the user's role to the same role.
 pub async fn set_role(
     jar: &RequestJar,
-    group_id: usize,
-    user_id: usize,
-    role_id: usize,
+    group_id: u32,
+    user_id: u32,
+    role_id: u32,
 ) -> Result<(), Box<Error>> {
     let url = format!(
         "https://groups.roblox.com/v1/groups/{}/users/{}",
@@ -69,9 +65,9 @@ pub async fn set_role(
 /// - 26: You cannot change the user's role to the same role.
 pub async fn set_rank(
     jar: &RequestJar,
-    group_id: usize,
-    user_id: usize,
-    role_id: usize,
+    group_id: u32,
+    user_id: u32,
+    role_id: u32,
 ) -> Result<(), Box<Error>> {
     set_role(jar, group_id, user_id, role_id).await?;
     Ok(())
@@ -91,9 +87,9 @@ pub async fn set_rank(
 /// - 200: The user's rank was not found.
 pub async fn modify_rank_by_amount(
     jar: &RequestJar,
-    group_id: usize,
-    user_id: usize,
-    amount: usize,
+    group_id: u32,
+    user_id: u32,
+    amount: u32,
 ) -> Result<(), Box<Error>> {
     let ranks = roles(jar, group_id).await?;
     let user_rank = user_role(jar, group_id, user_id).await?;
@@ -109,7 +105,7 @@ pub async fn modify_rank_by_amount(
     }
 
     // Get the index of the new rank
-    let new_rank_index = user_rank_index.unwrap() + amount;
+    let new_rank_index = user_rank_index.unwrap() + amount as usize;
     if new_rank_index >= ranks.len() {
         return Err(Box::new(Error::RobloxError(RobloxError {
             code: 2,
@@ -131,7 +127,7 @@ pub async fn modify_rank_by_amount(
 /// - 18: The operation is temporarily unavailable. Please try again later.
 /// - 23: You cannot change your own role.
 /// - 26: You cannot change the user's role to the same role.
-pub async fn promote(jar: &RequestJar, group_id: usize, user_id: usize) -> Result<(), Box<Error>> {
+pub async fn promote(jar: &RequestJar, group_id: u32, user_id: u32) -> Result<(), Box<Error>> {
     modify_rank_by_amount(jar, group_id, user_id, 1).await?;
     Ok(())
 }
@@ -146,7 +142,7 @@ pub async fn promote(jar: &RequestJar, group_id: usize, user_id: usize) -> Resul
 /// - 18: The operation is temporarily unavailable. Please try again later.
 /// - 23: You cannot change your own role.
 /// - 26: You cannot change the user's role to the same role.
-pub async fn demote(jar: &RequestJar, group_id: usize, user_id: usize) -> Result<(), Box<Error>> {
+pub async fn demote(jar: &RequestJar, group_id: u32, user_id: u32) -> Result<(), Box<Error>> {
     modify_rank_by_amount(jar, group_id, user_id, 1).await?;
     Ok(())
 }
